requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.29.5
azure-identity==1.15.0
azure-search-documents==11.4.0b6
azure-storage-blob==12.19.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.5
dataclasses-json==0.6.1
distro==1.8.0
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
isodate==0.6.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.331
langsmith==0.0.60
Markdown==3.5.1
marshmallow==3.20.1
msal==1.25.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
oauthlib==3.2.2
openai==0.27.0
outcome==1.3.0.post0
packaging==23.2
portalocker==2.8.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.0
pywin32==306
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
wsproto==1.2.0
yarl==1.9.2
azure-cosmos==4.5.1
bs4